% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_webscrape_utils.R
\name{clean_names}
\alias{clean_names}
\title{Clean Names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{a character vector of names to clean}
}
\value{
a character vector of clean names
}
\description{
This function is used internally in get_historic_ratios and is calibrated
specifically to take a company name and turn it into something that has
meaning in regards to the MacroTrends website url
}
