% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_filtered_stocks.R
\name{get_filtered_stocks}
\alias{get_filtered_stocks}
\title{Get Filtered Stocks}
\usage{
get_filtered_stocks(
  geo = "geo_usa",
  ind = "ind_any",
  sec = "sec_any",
  exchange = "exch_any",
  cap = "cap_any",
  fa_div = "any",
  fa_pe = "any",
  fa_peg = "any",
  fa_ps = "any",
  fa_pb = "any",
  fa_pc = "any",
  fa_eps5years = "any",
  fa_sales5years = "any",
  fa_debteq = "any",
  fa_payoutratio = "any",
  all_data = TRUE
)
}
\arguments{
\item{exchange}{}

\item{cap}{}

\item{fa_div}{}

\item{fa_pe}{}

\item{fa_peg}{}

\item{fa_ps}{}

\item{fa_pb}{}

\item{fa_pc}{}

\item{fa_eps5years}{}

\item{fa_sales5years}{}

\item{fa_debteq}{}

\item{fa_payoutratio}{}

\item{all_data}{}
}
\value{

}
\description{
Use FinViz.com to filter stocks and get a list of stocks that
meet some certain set of specifications.  Can return a dataframe
with lots of data, or just the stocks and their names
}
