% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_historical.R
\name{get_historical_dividends}
\alias{get_historical_dividends}
\title{Get Historical Dividends}
\usage{
get_historical_dividends(ticker, start_date = NULL, verbose = TRUE)
}
\arguments{
\item{ticker}{A string identifying the ticker of the stock to get 
information on, all caps}

\item{start_date}{A string identifying the date on which the information 
should start (ymd format), if NULL, returns all available information}

\item{verbose}{Logical to determine if the function will output info
as it runs}
}
\value{
a data frame with the following columns
  \itemize{
    \item{date = the date the dividend was paid}
    \item{amount_per_share = the amount per share in dollars}
    \item{ticker = the ticker of the stock/fund}
  }
}
\description{
A function to get historical dividend information from 
dividendinformation.com
}
