% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_historical.R
\name{get_historical_ratios}
\alias{get_historical_ratios}
\title{Get Historical Ratios}
\usage{
get_historical_ratios(
  tickers,
  ratios = c("pe-ratio", "price-sales", "price-book", "price-fcf", "debt-equity-ratio")
)
}
\arguments{
\item{tickers}{a vector of tickers representing the companies we want to
see.  They should be all uppercase string values}

\item{ratios}{a character vector identifying the ratios to use.  Can
take on the values "pe-ratio", "price-sales", "price-book", "price-fcf",
"debt-equity-ratio".  Defaults to all of them.  For now it is necessary 
to include the "pe-ratio" in all queries so that the date gets returned}
}
\value{
a data frame detailing historical ratio data for each of the stocks
  passed to the function via `tickers` and `names`
}
\description{
A function to scrape historical ratios from MacroTrends.com
}
