% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_webscrape_utils.R
\name{string_to_numeric}
\alias{string_to_numeric}
\title{String to Dollar}
\usage{
string_to_numeric(x)
}
\arguments{
\item{x}{a vector of character strings that represent abbreviated
dollar amounts}
}
\value{
a double vector of the same length as the input vector
}
\description{
Converts a numeric string (such as '53.5B') to its corresponding 
numeric value (eg. 53500000000)
}
