% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_historical.R
\name{get_historical_prices}
\alias{get_historical_prices}
\title{Get Historical Prices}
\usage{
get_historical_prices(tickers)
}
\arguments{
\item{tickers}{the tickers of the stocks that you want to retrieve historical prices for}
}
\value{
a data frame containing historical price information
}
\description{
A function to get the historical prices for a character vector of tickers
}
