% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_historical.R
\name{get_historical_prices}
\alias{get_historical_prices}
\title{Get Historical Prices}
\usage{
get_historical_prices(ticker, type = "stocks", years = 10, verbose = TRUE)
}
\arguments{
\item{ticker}{a string, the ticker for the company you wish to see data}

\item{type}{a string specifying the type of ticker you have}

\item{years}{an integer, how many years back you want the data to go (must
be 10 or less)}

\item{verbose}{logical specifying if the function should be chatty}
}
\value{
a data.frame containing historical information about the stock
}
\description{
Retrieves the historical prices for stocks from Nasdaq.com
}
\details{
Further specifications regarding the `type` parameter.
\itemize{
  \item{if stock, `type` = "stocks"}
  \item{if ETF, `type` = "etf"}
  \item{if mutual fund, `type` = "mutualfunds"}
}
}
